+++ presentation/src/main/java/com/danbam/presentation/app/BaseActivity.kt
@@ -12,13 +12,13 @@ import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.danbam.design_system.IndiStrawTheme
-import com.danbam.presentation.navigation.MainNavigationItem
-import com.danbam.presentation.navigation.authGraph
-import com.danbam.presentation.navigation.mainGraph
-import com.danbam.presentation.navigation.movieGraph
-import com.danbam.presentation.navigation.profileGraph
-import com.danbam.presentation.navigation.signUpGraph
-import com.danbam.presentation.ui.signup.SignUpViewModel
+import com.danbam.presentation.ui.main.navigation.MainNavigationItem
+import com.danbam.presentation.ui.auth.navigation.authGraph
+import com.danbam.presentation.ui.main.navigation.mainGraph
+import com.danbam.presentation.ui.movie.navigation.movieGraph
+import com.danbam.presentation.ui.profile.navigation.profileGraph
+import com.danbam.presentation.ui.auth.navigation.signUpGraph
+import com.danbam.presentation.ui.auth.signup.SignUpViewModel
 import com.google.accompanist.navigation.animation.AnimatedNavHost
 import com.google.accompanist.navigation.animation.rememberAnimatedNavController
 import dagger.hilt.android.AndroidEntryPoint
