+++ presentation/src/main/java/com/danbam/presentation/ui/auth/certificate/CertificateContract.kt
@@ -1,0 +1,18 @@
+package com.danbam.presentation.ui.auth.certificate
+
+data class CertificateState(
+    val phoneNumber: String = "",
+)
+
+sealed class CertificateSideEffect {
+    object EmptyPhoneNumberException : CertificateSideEffect()
+    object MatchPhoneNumberException : CertificateSideEffect()
+    object EnrollPhoneNumberException : CertificateSideEffect()
+    object NotEnrollPhoneNumberException : CertificateSideEffect()
+    object TooManyRequestPhoneNumberException : CertificateSideEffect()
+    object EmptyCertificateNumberException : CertificateSideEffect()
+    object WrongCertificateNumberException : CertificateSideEffect()
+    object ExpiredCertificateNumberException : CertificateSideEffect()
+    object TooManyRequestCertificateNumberException : CertificateSideEffect()
+    object SuccessCertificate : CertificateSideEffect()
+}
