+++ presentation/src/main/java/com/danbam/presentation/ui/auth/find/find_id/FindIdContract.kt
@@ -1,0 +1,5 @@
+package com.danbam.presentation.ui.auth.find.find_id
+
+data class FindIdState(
+    val id: String = "",
+)
