--- presentation/src/main/java/com/danbam/presentation/ui/certificate/CertificateContract.kt
@@ -1,18 +1,0 @@
-package com.danbam.presentation.ui.certificate
-
-data class CertificateState(
-    val phoneNumber: String = "",
-)
-
-sealed class CertificateSideEffect {
-    object EmptyPhoneNumberException : CertificateSideEffect()
-    object MatchPhoneNumberException : CertificateSideEffect()
-    object EnrollPhoneNumberException : CertificateSideEffect()
-    object NotEnrollPhoneNumberException : CertificateSideEffect()
-    object TooManyRequestPhoneNumberException : CertificateSideEffect()
-    object EmptyCertificateNumberException : CertificateSideEffect()
-    object WrongCertificateNumberException : CertificateSideEffect()
-    object ExpiredCertificateNumberException : CertificateSideEffect()
-    object TooManyRequestCertificateNumberException : CertificateSideEffect()
-    object SuccessCertificate : CertificateSideEffect()
-}
