--- presentation/src/main/java/com/danbam/presentation/ui/certificate/CertificateViewModel.kt
@@ -1,81 +1,0 @@
-package com.danbam.presentation.ui.certificate
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.danbam.domain.usecase.auth.CheckCertificateNumberUseCase
-import com.danbam.domain.usecase.auth.CheckPhoneNumberUseCase
-import com.danbam.domain.usecase.auth.SendCertificateNumberUseCase
-import com.danbam.presentation.util.android.errorHandling
-import com.danbam.presentation.util.parser.isPhoneNumber
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
-import org.orbitmvi.orbit.ContainerHost
-import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.postSideEffect
-import org.orbitmvi.orbit.syntax.simple.reduce
-import org.orbitmvi.orbit.viewmodel.container
-import javax.inject.Inject
-
-@HiltViewModel
-class CertificateViewModel @Inject constructor(
-    private val checkPhoneNumberUseCase: CheckPhoneNumberUseCase,
-    private val sendCertificateNumberUseCase: SendCertificateNumberUseCase,
-    private val checkCertificateNumberUseCase: CheckCertificateNumberUseCase,
-) : ContainerHost<CertificateState, CertificateSideEffect>, ViewModel() {
-    override val container = container<CertificateState, CertificateSideEffect>(CertificateState())
-
-    fun checkPhoneNumber(phoneNumber: String, type: String) = intent {
-        if (phoneNumber.isEmpty()) postSideEffect(CertificateSideEffect.EmptyPhoneNumberException)
-        else if (!phoneNumber.isPhoneNumber()) postSideEffect(CertificateSideEffect.MatchPhoneNumberException)
-        else {
-            viewModelScope.launch {
-                checkPhoneNumberUseCase(phoneNumber = phoneNumber, type = type).onFailure {
-                    it.errorHandling(unknownAction = {}, notFoundException = {
-                        postSideEffect(CertificateSideEffect.NotEnrollPhoneNumberException)
-                    }, conflictException = {
-                        postSideEffect(CertificateSideEffect.EnrollPhoneNumberException)
-                    }, tooManyRequestException = {
-                        postSideEffect(CertificateSideEffect.TooManyRequestPhoneNumberException)
-                    }, noContentException = {
-                        sendCertificateNumber(phoneNumber = phoneNumber)
-                    })
-                }
-            }
-        }
-    }
-
-    private fun sendCertificateNumber(phoneNumber: String) = intent {
-        viewModelScope.launch {
-            sendCertificateNumberUseCase(phoneNumber = phoneNumber).onFailure {
-                it.errorHandling(unknownAction = {}, tooManyRequestException = {
-                    postSideEffect(CertificateSideEffect.TooManyRequestCertificateNumberException)
-                }, noContentException = {
-                    reduce { state.copy(phoneNumber = phoneNumber) }
-                })
-            }
-        }
-    }
-
-    fun checkCertificateNumber(authCode: String) = intent {
-        if (authCode.isEmpty()) postSideEffect(CertificateSideEffect.EmptyCertificateNumberException)
-        else {
-            viewModelScope.launch {
-                checkCertificateNumberUseCase(
-                    authCode = authCode.toInt(),
-                    phoneNumber = state.phoneNumber
-                ).onFailure {
-                    it.errorHandling(unknownAction = {}, wrongDataException = {
-                        postSideEffect(CertificateSideEffect.WrongCertificateNumberException)
-                    }, noContentException = {
-                        postSideEffect(CertificateSideEffect.SuccessCertificate)
-                        reduce { state.copy(phoneNumber = "") }
-                    })
-                }
-            }
-        }
-    }
-
-    fun expiredCertificateNumber() = intent {
-        postSideEffect(CertificateSideEffect.ExpiredCertificateNumberException)
-    }
-}
