--- presentation/src/main/java/com/danbam/presentation/ui/find/find_id/FindIdContract.kt
@@ -1,5 +1,0 @@
-package com.danbam.presentation.ui.find.find_id
-
-data class FindIdState(
-    val id: String = "",
-)
