--- presentation/src/main/java/com/danbam/presentation/ui/find/find_password/FindPasswordContract.kt
@@ -1,11 +1,0 @@
-package com.danbam.presentation.ui.find.find_password
-
-sealed class FindPasswordSideEffect {
-    object EmptyPasswordException : FindPasswordSideEffect()
-    object EmptyRePasswordException : FindPasswordSideEffect()
-    object DifferentPasswordException : FindPasswordSideEffect()
-    object LengthPasswordException : FindPasswordSideEffect()
-    object MatchPasswordException : FindPasswordSideEffect()
-    object FailChangeException : FindPasswordSideEffect()
-    object SuccessChange : FindPasswordSideEffect()
-}
