--- presentation/src/main/java/com/danbam/presentation/ui/intro/IntroContract.kt
@@ -1,9 +1,0 @@
-package com.danbam.presentation.ui.intro
-
-data class IntroState(
-    val isNeedLogin: Boolean = false,
-)
-
-sealed class IntroSideEffect {
-    object SuccessLogin : IntroSideEffect()
-}
