--- presentation/src/main/java/com/danbam/presentation/ui/login/LoginContract.kt
@@ -1,11 +1,0 @@
-package com.danbam.presentation.ui.login
-
-
-sealed class LoginSideEffect {
-    object LoginSuccess : LoginSideEffect()
-    object EmptyIdException : LoginSideEffect()
-    object EmptyPasswordException : LoginSideEffect()
-    object MatchIdException : LoginSideEffect()
-    object MatchPasswordException : LoginSideEffect()
-    object UnKnownException : LoginSideEffect()
-}
