+++ presentation/src/main/java/com/danbam/presentation/ui/main/intro/IntroContract.kt
@@ -1,0 +1,9 @@
+package com.danbam.presentation.ui.main.intro
+
+data class IntroState(
+    val isNeedLogin: Boolean = false,
+)
+
+sealed class IntroSideEffect {
+    object SuccessLogin : IntroSideEffect()
+}
